Move ConjureLockRequestTest to conjure module
