+++ timelock-api/build.gradle
@@ -7,6 +7,11 @@ dependencies {
     compileOnly 'com.fasterxml.jackson.core:jackson-databind'
     compileOnly 'org.immutables:value::annotations'
     annotationProcessor 'org.immutables:value'
+
+    testImplementation project(':atlasdb-api')
+    testImplementation 'com.palantir.conjure.java.runtime:conjure-java-jackson-serialization'
+    testImplementation 'junit:junit'
+    testImplementation 'org.assertj:assertj-core'
 }
 
 subprojects {
